+++ Assets/Scripts/UI/HealthBar.cs
@@ -42,7 +42,7 @@ public class HealthBar : MonoBehaviour
             _currentDelayTime = delayTimer;
         }
     }
-    public void SetSliderMaxValue(int maxHealth)
+    public void SetSliderMaxValue(float maxHealth)
     {
         healthSlider.maxValue = maxHealth;
         delaySlider.maxValue = maxHealth;
@@ -50,7 +50,7 @@ public class HealthBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-     public void UpdateHealthUI(int currentHealth)
+     public void UpdateHealthUI(float currentHealth)
     {
         sliderTextHealth.text = (currentHealth).ToString();
         healthSlider.value = currentHealth;
